remove in-repo .hdr files
This type of asset should live outside of the repo, since churn on these
kinds of assets will inflate the size of the repo. In previous commits I
removed the links to the local .hdr files and instead used URLs to load
skyboxes. This commit just removes the .hdr files within the repo, since
they're no longer used, and keeping them in the repo could be confusing

I've rehosted the skyboxes to the following URL:
https://mush.network/files/sky

And the *.hdr files now live on the Osaka server in the following
directory:
/home/me/public/sky

Which you can log into, if you have ssh access, with this command:
ssh me@mush.network